@@ -1,4 +1,3 @@
-# encoding: UTF-8
 # This file is auto-generated from the current state of the database. Instead
 # of editing this file, please use the migrations feature of Active Record to
 # incrementally modify your database, and then regenerate this schema definition.
@@ -11,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 20170312123257) do
+ActiveRecord::Schema.define(version: 20170716113643) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -25,11 +24,10 @@ ActiveRecord::Schema.define(version: 20170312123257) do
     t.string   "author_type"
     t.datetime "created_at",    null: false
     t.datetime "updated_at",    null: false
+    t.index ["author_type", "author_id"], name: "index_active_admin_comments_on_author_type_and_author_id", using: :btree
+    t.index ["namespace"], name: "index_active_admin_comments_on_namespace", using: :btree
+    t.index ["resource_type", "resource_id"], name: "index_active_admin_comments_on_resource_type_and_resource_id", using: :btree
   end
-
-  add_index "active_admin_comments", ["author_type", "author_id"], name: "index_active_admin_comments_on_author_type_and_author_id", using: :btree
-  add_index "active_admin_comments", ["namespace"], name: "index_active_admin_comments_on_namespace", using: :btree
-  add_index "active_admin_comments", ["resource_type", "resource_id"], name: "index_active_admin_comments_on_resource_type_and_resource_id", using: :btree
 
   create_table "addresses", force: :cascade do |t|
     t.text     "full_address"
@@ -37,9 +35,8 @@ ActiveRecord::Schema.define(version: 20170312123257) do
     t.integer  "user_id"
     t.datetime "created_at",   null: false
     t.datetime "updated_at",   null: false
+    t.index ["user_id"], name: "index_addresses_on_user_id", using: :btree
   end
-
-  add_index "addresses", ["user_id"], name: "index_addresses_on_user_id", using: :btree
 
   create_table "admin_users", force: :cascade do |t|
     t.string   "email",                  default: "", null: false
@@ -54,10 +51,9 @@ ActiveRecord::Schema.define(version: 20170312123257) do
     t.inet     "last_sign_in_ip"
     t.datetime "created_at",                          null: false
     t.datetime "updated_at",                          null: false
+    t.index ["email"], name: "index_admin_users_on_email", unique: true, using: :btree
+    t.index ["reset_password_token"], name: "index_admin_users_on_reset_password_token", unique: true, using: :btree
   end
-
-  add_index "admin_users", ["email"], name: "index_admin_users_on_email", unique: true, using: :btree
-  add_index "admin_users", ["reset_password_token"], name: "index_admin_users_on_reset_password_token", unique: true, using: :btree
 
   create_table "coupons", force: :cascade do |t|
     t.string   "code"
@@ -87,12 +83,11 @@ ActiveRecord::Schema.define(version: 20170312123257) do
     t.string   "name"
     t.string   "phone_no"
     t.integer  "coupon_id"
+    t.index ["coupon_id"], name: "index_orders_on_coupon_id", using: :btree
+    t.index ["service_id"], name: "index_orders_on_service_id", using: :btree
+    t.index ["status_id"], name: "index_orders_on_status_id", using: :btree
+    t.index ["user_id"], name: "index_orders_on_user_id", using: :btree
   end
-
-  add_index "orders", ["coupon_id"], name: "index_orders_on_coupon_id", using: :btree
-  add_index "orders", ["service_id"], name: "index_orders_on_service_id", using: :btree
-  add_index "orders", ["status_id"], name: "index_orders_on_status_id", using: :btree
-  add_index "orders", ["user_id"], name: "index_orders_on_user_id", using: :btree
 
   create_table "pricings", force: :cascade do |t|
     t.integer  "price"
@@ -100,9 +95,8 @@ ActiveRecord::Schema.define(version: 20170312123257) do
     t.string   "pricable_type"
     t.datetime "created_at",    null: false
     t.datetime "updated_at",    null: false
+    t.index ["pricable_type", "pricable_id"], name: "index_pricings_on_pricable_type_and_pricable_id", using: :btree
   end
-
-  add_index "pricings", ["pricable_type", "pricable_id"], name: "index_pricings_on_pricable_type_and_pricable_id", using: :btree
 
   create_table "services", force: :cascade do |t|
     t.string   "name"
@@ -140,11 +134,11 @@ ActiveRecord::Schema.define(version: 20170312123257) do
     t.boolean  "terms_of_service"
     t.text     "coupons_ready",                     default: [],              array: true
     t.text     "coupons_used",                      default: [],              array: true
+    t.string   "auth_token"
+    t.index ["address_id"], name: "index_users_on_address_id", using: :btree
+    t.index ["email"], name: "index_users_on_email", unique: true, using: :btree
+    t.index ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
   end
-
-  add_index "users", ["address_id"], name: "index_users_on_address_id", using: :btree
-  add_index "users", ["email"], name: "index_users_on_email", unique: true, using: :btree
-  add_index "users", ["reset_password_token"], name: "index_users_on_reset_password_token", unique: true, using: :btree
 
   create_table "wardrobes", force: :cascade do |t|
     t.string   "name"
